Add storage window
It is tempting to add a new state so that an "issued" batch can move
into, say, "oblivion" ("expired" would be confusing.) That would however
distract from the core logic.

We still require CAs to maintain a list of all batches and its
assertions, but don't require them to expose them after the storage
window. I think that's ok.

GH #2